ingest/requirements.txt
requests
beautifulsoup4
qdrant-client
sentence-transformers
tqdm
python-dotenv
